+++ noodles-cram/src/async/io/reader/builder.rs
@@ -1,6 +1,5 @@
 use std::path::Path;
 
-use bytes::BytesMut;
 use noodles_fasta as fasta;
 use tokio::{
     fs::File,
@@ -74,7 +73,6 @@ impl Builder {
         Reader {
             inner: reader,
             reference_sequence_repository: self.reference_sequence_repository,
-            buf: BytesMut::new(),
         }
     }
 }